flask/requirements.txt
alembic==1.4.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.4
cffi==1.14.0
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ffmpeg==1.4
Flask==1.1.1
Flask-JWT==0.3.2
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
Mako==1.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numba==0.48.0
numpy==1.18.2
opencv-python==4.2.0.34
packaging==19.0
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
py==1.8.0
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.4.2
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
PySoundFile==0.9.0.post1
pytesseract==0.3.3
pytest==5.0.1
python-dateutil==2.8.1
python-editor==1.0.4
pytube3==9.6.4
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.6
terminado==0.8.3
tesseract==0.1.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
